backend/freshness_detection/requirements.txt
flask
flask-restx
werkzeug
opencv-python
numpy
Pillow
tensorflow
ultralytics

